Project structure global refactor add Spring boot framework
