+++ src/main/java/com/kafeshka/KafeshkaRS/KafeshkaRsApplication.java
@@ -1,0 +1,13 @@
+package com.kafeshka.KafeshkaRS;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class KafeshkaRsApplication {
+    public static void main(String[] args) {
+        SpringApplication.run(KafeshkaRsApplication.class, args);
+    }
+}
+
+
