+++ src/main/java/com/kafeshka/KafeshkaRS/customer/Customer.java
@@ -1,4 +1,4 @@
-package com.kafeshka;
+package com.kafeshka.KafeshkaRS.customer;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
