+++ src/main/java/com/kafeshka/KafeshkaRS/services/DiscountManagerService.java
@@ -1,8 +1,8 @@
-package com.kafeshka.services;
+package com.kafeshka.KafeshkaRS.services;
 
-import com.kafeshka.discount.Discount;
-import com.kafeshka.discount.DiscountManager;
-import com.kafeshka.order.Order;
+import com.kafeshka.KafeshkaRS.discount.Discount;
+import com.kafeshka.KafeshkaRS.discount.DiscountManager;
+import com.kafeshka.KafeshkaRS.order.Order;
 
 import java.util.List;
