--- src/main/java/com/kafeshka/KafeshkaRun.java
@@ -1,8 +1,0 @@
-package com.kafeshka;
-
-public class KafeshkaRun {
-    public static void main(String[] args) {
-
-
-    }
-}
