+++ src/main/resources/application.properties
@@ -1,0 +1,32 @@
+# Application Name
+spring.application.name=Kafeshka RS
+
+# Server Port
+server.port=8082
+
+# Disable Resource Caching
+spring.web.resources.cache.cachecontrol.max-age=0
+
+# Enable Thymeleaf
+spring.thymeleaf.enabled=true
+
+# Multipart File Upload Configuration
+spring.servlet.multipart.enabled=true
+spring.servlet.multipart.file-size-threshold=2KB
+spring.servlet.multipart.max-file-size=10MB
+spring.servlet.multipart.max-request-size=10MB
+spring.servlet.multipart.location=uploads
+
+# MySQL DataSource Configuration
+#spring.datasource.url=jdbc:mysql://localhost:3386/empoweringhr?useSSL=false
+#spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
+#spring.datasource.username=empower
+#spring.datasource.password=test123
+
+# Hibernate Configuration
+#spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.MariaDB103Dialect
+##spring.jpa.hibernate.ddl-auto=create-drop
+#spring.jpa.hibernate.ddl-auto=update
+#spring.jpa.generate-ddl=true
+#spring.session.store-type=jdbc
+#spring.session.jdbc.initialize-schema=always
