+++ src/test/java/com/kafeshka/KafeshkaRS/KafeshkaRsApplicationTest.java
@@ -1,0 +1,14 @@
+package com.kafeshka.KafeshkaRS;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+@SpringBootTest
+class KafeshkaRsApplicationTest {
+
+    @Test
+    void contextLoads() {
+    }
+}
